Swipe directory modified
