+++ src/test/java/swipe/SwiperRequestModelTest.java
@@ -1,0 +1,34 @@
+package swipe;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class SwiperRequestModelTest {
+
+    @BeforeEach
+    void setUp() {
+    }
+
+    @AfterEach
+    void tearDown() {
+    }
+
+    @Test
+    void getAccepted() {
+    }
+
+    @Test
+    void setAccepted() {
+    }
+
+    @Test
+    void getPotential() {
+    }
+
+    @Test
+    void setPotential() {
+    }
+}
