+++ src/test/java/swipe/SwiperResponseModelTest.java
@@ -1,10 +1,11 @@
+package swipe;
+
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class SwiperControllerTest {
+class SwiperResponseModelTest {
 
     @BeforeEach
     void setUp() {
@@ -12,9 +13,5 @@ class SwiperControllerTest {
 
     @AfterEach
     void tearDown() {
-    }
-
-    @Test
-    void create() {
     }
 }
