+++ src/test/java/swipe/screen/SwiperPresenterFormatterTest.java
@@ -1,3 +1,5 @@
+package swipe.screen;
+
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
